requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openpyxl==3.1.2
aiofiles==23.2.1 